src/requirements.txt
Jinja2==3.1.6
Markdown==3.8.2
MarkupSafe==3.0.2
PyYAML==6.0.2
requests
